本次提交内容：第二个art串口尝试；环岛转弯代码框架（待验证）

+++ new_project/new_peoject/code/communication.h
@@ -33,7 +33,7 @@
 #define MOTORCYCLE 15
 
 extern int now_distance_x;
-extern unsigned long now_distance_y;
+extern unsigned int now_distance_y;
 
 void My_Communication_Init(void);
 void uart1_rx_interrupt_handler(void);
